requirements.txt
django
requests
djangorestframework
numpy
joblib
scikit-learn
numpy
pandas
streamlit
